Rearrange code to decrease some of the file sizes:
1. Split single_test_basic into single_test_array and single_test_scalar
2. Move mmio32_register_layouts and mmio64_register_layouts into the mmio module.

Also, adds #![no_std] to tests where possible.
