+++ src/layouts.rs
@@ -306,51 +306,3 @@ macro_rules! register_layouts {
         $crate::internal::register_layouts!($crate $($arguments)*);
     }
 }
-
-/// An alias for [`register_layouts!`] with `#![buses(Mmio32)]` at the top. In other words:
-/// ```
-/// use tock_registers::{mmio32_register_layouts, Mmio32};
-/// mmio32_register_layouts! {
-///     // Register definitions here
-/// }
-/// ```
-/// is equivalent to:
-/// ```
-/// use tock_registers::{register_layouts, Mmio32};
-/// register_layouts! {
-///     #![buses(Mmio32)]
-///     // Register definitions here
-/// }
-/// ```
-// Note: Anytime this is changed, expand_macros should be updated as well, as it has to duplicate
-// this definition!
-#[macro_export]
-macro_rules! mmio32_register_layouts {
-    {$($arguments:tt)*} => {
-        $crate::internal::register_layouts!($crate #![bus($crate::Mmio32)] $($arguments)*);
-    }
-}
-
-/// An alias for [`register_layouts!] with `#![buses(Mmio64)]` at the top. In other words:
-/// ```
-/// use tock_registers::{mmio64_register_layouts, Mmio64};
-/// mmio64_register_layouts! {
-///     // Register definitions here
-/// }
-/// ```
-/// is equivalent to:
-/// ```
-/// use tock_registers::{register_layouts, Mmio64};
-/// register_layouts! {
-///     #![buses(Mmio64)]
-///     // Register definitions here
-/// }
-/// ```
-// Note: Anytime this is changed, expand_macros should be updated as well, as it has to duplicate
-// this definition!
-#[macro_export]
-macro_rules! mmio64_register_layouts {
-    {$($arguments:tt)*} => {
-        $crate::internal::register_layouts!($crate #![bus($crate::Mmio64)] $($arguments)*);
-    }
-}
