+++ tests/all_block_fields.rs
@@ -5,6 +5,8 @@
 
 //! Tests a register block with each type of field.
 
+#![no_std]
+
 // Note that we don't use-import the macro because this file is also used for expand_macros_test,
 // which builds the expanded code without the proc_macros feature.
 use tock_registers::{Read, Write};
